fix: resolve clippy warnings with -D warnings
- Add #[allow(dead_code)] for unused structs and methods reserved for future use
- Add #[allow(clippy::too_many_arguments)] for constructors with many params
- Add KpiHistory type alias to reduce type complexity
- Fix if-same-then-else in virustotal verdict calculation
- Run cargo fmt to fix formatting

+++ crates/tw-connectors/src/http.rs
@@ -420,7 +420,7 @@ fn rand_jitter() -> Duration {
     use std::hash::{Hash, Hasher};
     let mut hasher = std::collections::hash_map::DefaultHasher::new();
     std::time::Instant::now().hash(&mut hasher);
-    let jitter_ms = (hasher.finish() % 100) as u64;
+    let jitter_ms = hasher.finish() % 100;
     Duration::from_millis(jitter_ms)
 }
